Implement alignment 'inheritance'
Previously, there was no way for a function taking an 8-byte aligned
vector to accept a 16-byte aligned vector, whilst it should (since it'd
also be 8-byte aligned as well, obviously).

This commit changes things a bit so this is possible now (at least for
power-of-two alignment values, which should suit our needs).

Thanks to Gábor Lehel on `Haskell Cafe`_.

.. _Haskell Cafe: http://www.haskell.org/pipermail/haskell-cafe/2012-July/102192.html