Make sure we only create vectors of suitable length
Whilst this change might be somewhat controversial, I want to assure
all memory chunks we allocate for vectors are (in bytes) properly
aligned at the end (i.e. a 16-byte (128 bit) aligned vector will always
be a chunk of memory of N * 16 bytes long).

For now, `error` is used in the allocation function if this requirement
is not met.